backend/requirements.txt
django-rest-framework
django-storages
django-channels
django-cors-headers
pillow
